Column Ruleset code shift and SetAuth error handling.
Shifted code from ColumnUI.cs to SessionRuleSet.cs -- which is the best place for session logic to live.
Added error handling into the logic for SetAuthentication for the Office365 Auth inspector tab. We are doing lots of logic checks with the response body. Added code to make sure the expected responses are there before executing the code to ensure we don't run into null object references at runtime.