Further updates to SAML response parser
More code moved into repeatedly used functions.
Stopped update check on live trace again. Maybe this is just my corp machine, but I keep running into issues with it when the system proxy is set. Since the availability of a system update is now stored in an application preference there is now arguably no need to be checking on live trace anymore.