Session Times HML Markup.
Small updates to the SessionProcessor for Server Think Time and Transit Time.
Language updates to the tab page.
HTML markup for session timers server think time vs. transit time.
Renamed The element back to EXOFiddlerInspector in the update checker. This isn't worth changing just for branding, trying to touch the update mechanism as little as possible.
Minor updates to text in HTTP 403 for clearer language.