Update SessionProcessor.cs
The more I thought about the problem I am trying to solve the more I realised there couldn't be one check on a session. There actually needed to be three of them.
1. Authentication
2. Session Type
3. Response Server
For each of these there is now a confidence level Get/Set so as a session progresses through the logic in the SessionProcessor as much code as possible can be eliminated and the most accurate results are produced.

Still working through my lab traces for various scenarios to dial this new system in, but it's now much easier to control which of these 3 is set in stone based on session detections and which to allow logic further on down the line to run.