Added RuleSetVersion, branch URL and master URL into settings.
Coded these three settings into the RuleSetProcessor.
Design intention is the default value for rulsetVersion is 19000101, meaning the client is yet to or has never called out to the Github for an update on the ruleset. Only once new ruleset files have been pulled from the github and stored locally should this value be updated to that day's date. For example 20230428.