Minor improvements.
Renamed ExtensionEnabled to ExtensionSessionProcessingEnabled to better reflect what this Boolean does.
Renamed ExchangeVersion.json RulesetLastUpdated to NextUpdateCheck to better reflect the intention of this timestamp. It will be used to set as x hours or days in the future and check current timestamp against it to control how frequently web calls for updates are made.
Removed an unnecessary if then statement in MenuUI for MiProcessAllSessions.Enabled.