Fixes: ExtensionURLs & ExtensionVersion Json Reading.
As close as I can tell adding a try catch to the deserialize line in settingshandler.cs had the odd effect of fixing this issue for URLs and versions.
Even if this is not right, I don't really care too much, just needed this fixed.
Removed some references to ActivationService in some classes, only one is needed to mark sure the extension calls it once and starts everything running.
Updated Fiddler logging to show assembly name and class name.