Proof of concept compiling code at run time.
Hard coded a base64 string into function to perform CodeDom Compiler on.
Have the CodeDom compiler running and compiling the assembly.
SessionFlaghandler class needs to be public to allow this to work.
Working on getting the reflection coded up to run the code against a session and write session flags to the session.