Separating out ruleset to it's own DLL.
To do this I needed to add another project within the solution, so I needed to go back to VS2019 and .Netframework 4.8.
Clean up / removal on ruleset code and Json files in main application project.
SessionHandler now calls out to RunRuleset in the ruleset assembly.
This opens the door to exploring how ruleset updates can be delivered outside of application version updates.