Update RuleSetHandler.cs
Coded up removing assemblies which are not the latest.
As I suspected running into file in use conflicts. Can't delete a loaded assembly and can't unload a loaded assembly without unloading the AppDomain, which from jasonsla is essentially needing to close Fiddler to unload.
Suspect accompanying PowerShell script might be the way to accomplish this goal.
PS:
- Close Fiddler.
- Identify and delete old ruleset dll files.
- reload Fiddler.