Response Body Bug Fix.
UtilFindInResponse has stopped working, since April 19th.
Using this point in time to compare project structure against to work out why this isn't working in the current project layout.
After comparing how the code structure I worked out somewhere I had switched to using utilFindInResponse for the HTTP 503 FederatedSTSUnreachable rule / function.
For some odd reason it just doesn't want to work for this type of session.
Created the SessionWordSearch utility function to perform the search and match.
Planning to change over all use of utilFindInResponse to this utility function as I am now not sure of the robustness of this built in function.