Introduced SimpleSessionAnalysis.cs.
Thinking through it, there are many session response codes which only really provide some different text on the session analysis.
Putting the different text in SessionClassification.json and removing a load of repeated code which would likely only introduce my human error.
This'll make a more robust code base, but does reply on the presence of the data in the json file / Fiddler setting.