Expanding out session classification Json data & performance gain.
Working through major session analysis logic and pushing more text into the Json file, so that less of this is hard coded into the extension.
Also updated LoadSaz / SaveSaz & SessionFlagAnalysis functions so that when a SAZ file is saved with the extension, the analysis is kept in the Microsoft365FiddlerExtensionJson session flag. When the Saz file is loaded again or session processing is reran the session processing calls the data from the session instead of going through the heavy lift of session logic again.
Massive performance gain for this situation, session colourisation and analysis are loaded as text instead of running through the session logic in the ruleset.
Coding in clear & recalculate session analysis from sessions.
Clear is effectively remove session colourisation.
Recalculate is removing any session analysis stored in the Microsoft365FiddlerExtensionJson session flag, and running through the ruleset to set the flags again.
Recalculate can be useful if a user wants to have the extension recalculate the session analysis on sessions. When multiple engineers are working on a trace, session severity could have been set manually in the context menu.
Next action is to pull session related items out of the main menu.