Microsoft365 URLs and IP addresses web service integration.
Minor fix to OWA notification channel session color.
Adding functionality to integrate with the Microsoft365 URLs and IP addresses web service.
Intending to be able to provide feedback on whether a given IP address is listed as a Microsoft365 IP address or not in the inspector.
This will be the start of expanding out more information on what service/endpoint/device responded to a given session in a trace.