Json validator.
Added functions to read Json data in HTTP 200 OK sessions.
If the Json passes the validator, we received all the data from the server, so mark that actually OK.
If the Json is incomplete / malformed it won't pass the validator, so mark that as session severity 50, concerning, black with red text.
Change cuts down on the number of sessions the extension flags with the HTTP 200 Lurking Errors function.