TLSVersion.
Created a dedicated class for the TLS version to be set within the extension's session flag.
Calculate this once per session and store it, so it can be used multiple times.
Needed this to work through the analysis being put together in ConsolidatedAnalysisReportService.cs
Got the CARS to the point where it exports to a HTML file via a content menu click.