Update HostIP.cs
Two fixes:
Handle a null or absent X-HostIP session flag.
Handle when networking service is called on a X-HostIP session flag.
Fixed a null object exception I saw when the flag isn't even present on the session. This might be an older issue with Fiddler and the old test traces I work with, but it's fixed in the extension either way.