Ruleset SessionClassification Json unavailable null object exception fixes.
Fixes to issues setting session classification and version json on a fresh machine.
Preventing ruleset running on sessions when the Session Classification preference is unavailable. This creates null object exception failures.
As soon as the Session Classification preference is created, usually after the first 5 or 6 sessions are processed after initial install, the ruleset runs there after.
This is a first run after install scenario.