Project clean up.
Commenting more functions.
Renamed some functions for simplicity.
Reorganized Preferences.cs so it makes more sense and is less of a mess.
Cleaned up some unused functions in classes.
Changed ResponseServer.cs; CDN was specifically looking for Akami. Probably the most likely for M365 cloud traffic, but might as well broaden the scope of the rule and let it for any CDN that responds with the same X header in the response.
Added opera as a browser executible in SessionType.cs.