v2.0.3: Language resource split.
Splitting out extension and ruleset language resource files.
This is a bug fix to eliminate the need for extension updates when the ruleset is updated. I realised this was an issue when recently creating a new rule. I noticed I was making changes for the new rule in the main extension DLL. This is a problem if the design intention is to make the code in the extension more stable / updated with much less frequency, and the majority of updates are pushed through the Ruleset DLL.

Goal is to end up with:
Office365FiddlerExtension DLL contains only application strings.
Office365FiddlerExtensionRuleset DLL contains only ruleset things.
Everything to add or update a rule lives within the Ruleset DLL.

Wanted to create a minimal set of services classes in the ruleset project, so calls back over to the main extension code are minimized. Means some duplicated code, but it is now fairly long standing code which won't ever likely change. Mental note to revisit and call through the services in the main extension project to minimise any duplicate code.