Building out Tab Page to be the primary control for the extension.
Removed the enable/disable menu option from the MenuUI.
Unfortunately driving enable/disable can't update UI controls in the tab page.
So the enable/disable control needs to live only on the tab page.
Added public functions to update UI controls when extension is enabled / disabled for MenuUI and ContextMenuUI.
Updated LoadSaz to be more intelligent when it comes to loading Saz files previously saved with the extension. The presence of extension session flags makes colourising sessions a non-intensive compute task, so allowing that to happen even when the extension is disabled. Saz files which need session processing are not allowed with the extension disabled.

To do: Remove sections from the About form.