WhoIsClient: Introduced WhoIs Lookups to show the organization name for the IP address.
Thinking I want a check box on the form to disable this though as there is a slight delay as the WHOIS server is located and the call on port 43 goes out and the response comes back.
Either that or perhaps someone is sitting behind a firewall and they can't access port 43.

Found some strings in Office365TabPage.cs which need converting over to LangHelper + To Do list.