Whois and NeverWebCall.
Coded out Whois so the user has the choice of running Whois queries when check IP addresses on the Tab Page in the extension.
Implemented improvements into the About and Tab Page UI controls to adhere to the Never Web Call preference. UI now clearly does not show any information if it is setup to run in a silo with Never Web Call enabled.