Bug fix: Multiple Message boxes around session threshold.
Prior versions of the extension threw a message box to confirm large session analysis.
This has gotten problematic with the number of ways multiple repeated prompts are generated.
So instead added a button on the tab page, and options in the menu and context menu to stop session analysis.
This gives the user a way to cancel off long running session processing if they want.